added some misc data structures + an asm implementation of fixed-size arrays + memcpy

+++ fp_algorithms/functional_list_algorithms.cpp
@@ -5,8 +5,8 @@
 
 template <class T>
 class RCBaseObject {
-    int  _rc = 0;
-    bool _released = false;
+    int16_t _rc = 0;
+    bool    _released = false;
 
     RCBaseObject () {}
     // virtual ~RCBaseObject () {}
